Es6 classes refactoring (#2)
* Rearrange methods

* Initiate exporter class. Move .save / .addMedia methods

* Move to class .addCard / .update

* Add test for .addCard
